refactor(tag): tag 조회 API 변경
