+++ backend/src/main/java/codezap/tag/repository/TemplateTagJpaRepository.java
@@ -1,4 +1,4 @@
-package codezap.template.repository;
+package codezap.tag.repository;
 
 import java.util.List;
