+++ backend/src/main/java/codezap/template/domain/TemplateTag.java
@@ -11,6 +11,7 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.MapsId;
 
 import codezap.global.auditing.BaseTimeEntity;
+import codezap.tag.domain.Tag;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
