+++ backend/src/main/java/codezap/template/dto/response/FindAllTemplatesResponse.java
@@ -3,8 +3,8 @@ package codezap.template.dto.response;
 import java.time.LocalDateTime;
 import java.util.List;
 
+import codezap.tag.domain.Tag;
 import codezap.template.domain.SourceCode;
-import codezap.template.domain.Tag;
 import codezap.template.domain.Template;
 import io.swagger.v3.oas.annotations.media.Schema;
