+++ backend/src/main/java/codezap/template/dto/response/FindTemplateResponse.java
@@ -4,8 +4,8 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 import codezap.category.dto.response.FindCategoryResponse;
+import codezap.tag.domain.Tag;
 import codezap.template.domain.SourceCode;
-import codezap.template.domain.Tag;
 import codezap.template.domain.Template;
 import io.swagger.v3.oas.annotations.media.Schema;
