+++ backend/src/test/java/codezap/template/controller/TemplateControllerTest.java
@@ -41,6 +41,7 @@ import codezap.member.dto.MemberDto;
 import codezap.member.repository.FakeMemberRepository;
 import codezap.member.repository.MemberRepository;
 import codezap.member.service.MemberService;
+import codezap.tag.service.TemplateTagService;
 import codezap.template.dto.request.CreateSourceCodeRequest;
 import codezap.template.dto.request.CreateTemplateRequest;
 import codezap.template.dto.request.UpdateSourceCodeRequest;
@@ -53,7 +54,6 @@ import codezap.template.repository.FakeThumbnailRepository;
 import codezap.template.repository.TemplateRepository;
 import codezap.template.service.SourceCodeService;
 import codezap.template.service.TemplateService;
-import codezap.template.service.TemplateTagService;
 import codezap.template.service.ThumbnailService;
 import codezap.template.service.facade.MemberTemplateApplicationService;
 import codezap.template.service.facade.TagTemplateApplicationService;
